- merge from usability-branch revision 4509 to head. - update version number to 3.2-SNAPSHOT
git-svn-id: https://svn.code.sf.net/p/openuss/code/branches/openuss-plexus-3.2@4528 60046477-0c10-0410-bf4b-e552a4ff1768

+++ plexus/plexus-web/src/main/java/org/openuss/web/search/ExtendedSearchPage.java
@@ -369,7 +369,7 @@ private static final Logger logger = Logger.getLogger(ExtendedSearchPage.class);
 		List<SelectItem> institutesToDisplay = new ArrayList<SelectItem>();
 		InstituteInfo instituteInfo;
 		
-		if(suborganisationId.intValue() > Constants.EXTENDED_SEARCH_GET_ALL.intValue()){
+		if(suborganisationId > Constants.EXTENDED_SEARCH_GET_ALL){
 			List<?> institutes = instituteService.findInstitutesByDepartmentAndEnabled(suborganisationId, true);
 			if(institutes.size() > 0){
 				institutesToDisplay.add(getAllSelectItem());
